requirements.txt
# Production requirements

boto3==1.34.139
celery==5.4.0
crispy-bootstrap3==2024.1
dj-database-url==2.2.0
Django==5.0.6
django-clear-cache==0.3
django-crispy-forms==2.2
django-dbtemplates==4.0
django-impersonate==1.9.4
django-markdown2==0.3.1
django-polymorphic==3.1.0
django-recaptcha==4.0.0
django-reversion==5.0.12
django-storages==1.14.3
djangorestframework==3.15.2
easy-thumbnails==2.8.5
gunicorn==22.0.0
psycopg2==2.9.9
pylibmc==1.6.3
PyPDF2==3.0.1
pytz==2024.1
redis==5.0.6
sentry-sdk[django]==2.8.0
weasyprint==62.3

# Testing requirements
beautifulsoup4==4.12.3
django-debug-toolbar==4.4.2
django-selenium-login==2.0.0
factory-boy==3.3.0
html5lib==1.1
pillow==10.3.0
selenium==4.22.0
tblib==3.0.0


# # Dependencies of direct requirements

amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
botocore==1.34.145
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
django-appconf==1.0.6
Faker==26.0.0
fonttools==4.53.1
h11==0.14.0
idna==3.7
jmespath==1.0.1
kombu==5.3.7
markdown2==2.5.0
outcome==1.3.0.post0
packaging==24.1
prompt_toolkit==3.0.47
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
s3transfer==0.10.2
setuptools==71.0.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.1
tinycss2==1.3.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
zopfli==0.2.3